feat(ffi verifier): move pactffi_verifier_cli_args to mod.rs, tidy, add docs
